% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype}
\alias{genotype}
\alias{is_genotype}
\title{Create a genotype object}
\usage{
genotype(Z, subject.include = NULL, subject.exclude = NULL,
  snp.include = NULL, snp.exclude = NULL)

is_genotype(x)
}
\arguments{
\item{Z}{A numeric genotype matrix (dosage matrix) or coercible with
as.matrix - rows correspond to individuals and columns correspond to SNPs.
Use 'NA' for missing values. The column names of this matrix should
correspond to SNP names in the \code{snpinfo} object. The row names of this
matrix should correspond to the id in the \code{phenotype} object.}

\item{subject.include}{(optional) character vector of the subjects in
\code{rownames(Z)} to be included in the analysis.  See Details.}

\item{subject.exclude}{(optional) character vector of the subjects in
\code{rownames(Z)} to be excluded in the analysis.  See Details.}

\item{snp.include}{(optional) character vector of the SNPs in
\code{colnames(Z)} to be included in the analysis.  See Details.}

\item{snp.exclude}{(optional) character vector of the SNPs in
  \code{colnames(Z)} to be excluded in the analysis.  See Details.}
}
\value{
an object of class 'genotype'.

  An object of class 'genotype' is a list containing at least the following
  components:

\itemize{
 \item data data frame of the data for analysis
 \item CAC calulated allele count.
 \item CAF calulated allele frequency.
 \item gender column name containing the gender of the subjects
 \item include the subjects in \code{data} that will be included is further analysis.
 \item exclude the subjects in \code{data} that will be excluded is further analysis.
}

It is important to note that return \code{include} and \code{exclude} are taken from data and not from the input parameters.
}
\description{
This function creates a genotype object.
}
\details{
Z should be a numeric matrix

  Typically either \code{subject.include} or \code{subject.exclude}, or both,
  is set to \code{NULL}. It is important to note that return \code{include}
  and \code{exclude} are taken from the subjects in \code{Z} and not from the
  input parameters.

  Typically either \code{snp.include} or \code{snp.exclude}, or both, is set
  to \code{NULL}. It is important to note that return \code{include} and
  \code{exclude} are taken from the snps in \code{Z} and not from the input
  parameters.
}

