% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwasr.R
\docType{package}
\name{gwasr}
\alias{gwasr}
\alias{gwasr-package}
\title{gwars: Provide a framework to make GWAS easier.}
\description{
A framework to make GWAS easier and more consistant.  It creates classes and functions to make common tasks standardized.
#'
To learn more about gwasr, start with the vignettes:
\code{browseVignettes(package = "gwasr")}
}

