% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan Plot}
\usage{
plot_manhattan(
  data = data,
  P = "P",
  CHR = "CHR",
  BP = "BP",
  maxP = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colrs = NULL,
  sigA = NULL,
  sigAcol = NULL,
  sigB = NULL,
  sigBcol = NULL,
  ylims = NULL,
  ybreaks = NULL
)
}
\arguments{
\item{data}{Data frame which contains GWAS summary statistics (REQUIRED)}

\item{P}{A column name which contains P values, default is "P", column should be numeric.}

\item{CHR}{A column name which contains chromosome ID, default is "CHR", column should be numeric or character.}

\item{BP}{A column name which contains base pair location, default is "BP", column should be numeric.}

\item{maxP}{A maximum P-value cutoff (sometimes, it's not a good use of memory to plot millions of p values above 0.1), default is 1.}

\item{xlab}{Optional x-axis label, default is NULL.}

\item{ylab}{Optional y-axis label, default is "-log10(P)"}

\item{main}{Optional plot title, default is "Manhattan plot"}

\item{colrs}{Optional character vector of colors, default is set by package.}

\item{sigA}{Optional adjustment to stringent p-value reference, default is 3e-8}

\item{sigAcol}{Option string to adjust the color of sigA line}

\item{sigB}{Optional adjustment to stringent p-value reference, default is 1e-5}

\item{sigBcol}{Option string to adjust the color of sigB line}

\item{ylims}{Optional adjustments to y limits, default adjusts it to data}

\item{ybreaks}{Optional adjsuutemts to y breaks, default adds whole numbers in increments of 1}
}
\value{
ggplot object
}
\description{
\emph{plot_manhattan} takes chromosome, base pair, and p-value data and creates a ggplot based Manhattan plot.
}
\details{
Basic functionality: \emph{plot_manhattan} takes chromosome and basepair address information, as well as p values from the GWAS summary statistics. Subsequently, data are additionally processed to generate Manhattan plot where chromosome and basepair information are used to generate x-axis, and the \eqn{-log_10 (P)} values are plotted on the y axis.
}
\examples{
plot_manhattan(ADHDMeta)
}
\author{
Franjo, Mingjing, and Xiaoxiao
}
