% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpinfo.R
\name{snpinfo}
\alias{is_snpinfo}
\alias{snpinfo}
\title{Create a snpinfo object}
\usage{
snpinfo(data, snpNames = "Name", aggregateBy = "gene", chr = NULL,
  pos = NULL, ref = NULL, alt = NULL, filterBy = NULL,
  filterFun = NULL, otherCols = NULL)

is_snpinfo(x)
}
\arguments{
\item{snpNames}{The field of si where the SNP identifiers are found. Default is 'Name'}

\item{aggregateBy}{The field of si on which the results were aggregated. Default is 'gene'. For single snps which are intended only for single variant analyses, it is reccomended that they have a unique identifier in this field.}

\item{chr}{(optional) The field of si where the chromosome identifiers are found. See Details.}

\item{pos}{(optional) The field of si where the snp position identifiers are found. See Details.}

\item{ref}{(optional) The field of si where the reference allele for the snps is found. See Details.}

\item{alt}{(optional) The field of si where the alternate allele for the snps is found. See Details.}

\item{filterBy}{(optional) The (typically logical) field in si in which to keep in further analysis. See Details}

\item{filterFun}{(optional) a function to apply to multiple '.filterBy' fields via \code{Reduce}. See Details.}

\item{otherCols}{(optional) other fields of si to propagate through. See Details.}

\item{si}{a data frame containing the SNP information.  Must contain fields given in '.snpNames' and '.aggregateBy'.}
}
\value{
an object of class 'snpinfo'.  This is a list typicaly used to feed
  into an analysis function.

  An object of class 'snpinfo' is a list containing at least the following
  components:

\itemize{
 \item data containing the SNP information
 \item snpNames column name containing the names of SNPs
 \item aggregateBy column name which results are to be aggregated
 \item chrCol column name containing the chromosome a SNP is found at.
 \item posCol column name containing the position in a chromosome a SNP is found at.
 \item refCol column name containing the reference allele.
 \item altCol column name containing the alternate allele.
}

It is important to note that return \code{include} and \code{exclude} are taken from data and not from the input parameters.
}
\description{
This function creates a snpinfo object.
}
\details{
It is HIGHLY recommended that \code{.chr}, \code{.pos}, \code{.ref}, and \code{.alt} be supplied.  For

\code{.chr} and \code{.pos} are used for branching and output in subsequent
analysis.  Useful for analysis where the X chromosome is handled differently
than other chromosomes.

\code{ref} and \code{alt} are used in calculating various SNP and subject
summary statistics in 'genotype' objects.

\code{.filterBy} and \code{.filterFun} are used to create filters, typically
at the meta-analysis stage (e.g., because they are intronic or common).
\code{TRUE} indicates that the SNP should be kept.  If \code{.filterFun} is
specified \code{Reduce} will be applied to the columns specified by
\code{.filterBy}.  Note \code{Reduce} is called the result must be a logical
vector for each entry in \code{si}. If \code{.filterFun} produces
\code{NA}'s, these will be changed to \code{FALSE} with a warning.

\code{.otherCols}
}
\seealso{
\code{Reduce}
}

